Enny Finish 2

+++ app/Http/Controllers/detail_rekam_medis.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Session;
 use App\DetailRekamMedis;
 use DB;
 use Alert;
@@ -37,17 +38,23 @@ class detail_rekam_medis extends Controller
      */
     public function store(Request $request)
     {
-        $data = [
-            'id_rm' => $request->id_rm,
-            'id_jadwal' => $request->id_jadwal,
-            'area_stimulasi' => $request->area_stimulasi,
-            'keterangan' => $request->keterangan
-        ];
+        if (Session::get('login')) {
+            
+            $data = [
+                'id_rm' => $request->id_rm,
+                'id_jadwal' => $request->id_jadwal,
+                'area_stimulasi' => $request->area_stimulasi,
+                'keterangan' => $request->keterangan
+            ];
 
-        DetailRekamMedis::insert($data);
+            DetailRekamMedis::insert($data);
 
-        Alert::success('Data berhasil ditambahkan')->autoclose(3500);
-        return redirect('detail_rekam_medis/'.$request->id_rm);
+            Alert::success('Data berhasil ditambahkan')->autoclose(3500);
+            return redirect('detail_rekam_medis/'.$request->id_rm);
+        }else{
+            Alert::error('Silahkan login terlebih dahulu!')->autoclose(3500);
+            return redirect('/login');
+        }
     }
 
     /**
@@ -58,13 +65,19 @@ class detail_rekam_medis extends Controller
      */
     public function show($id)
     {
-        $jadwal = DB::table('jadwal_terapis')->select('jadwal_terapis.*')->join('h_rekam_medis', 'jadwal_terapis.id_asses', '=', 'h_rekam_medis.id_asses')->where('h_rekam_medis.id_rm', $id)->whereNotIn('id_jadwal', DB::table('d_rekam_medis')->pluck('id_jadwal'))->where('status_terapis',  '=', 'Hadir')->get();
-        $detail = DetailRekamMedis::join('jadwal_terapis', 'd_rekam_medis.id_jadwal', '=', 'jadwal_terapis.id_jadwal')->where('id_rm', $id)->get();
-        return view('rekam_medis.detail_rekamedis', [
-            'detail' => $detail,
-            'id_rm' => $id,
-            'jadwal' => $jadwal
-        ]);
+        if (Session::get('login')) {
+            
+            $jadwal = DB::table('jadwal_terapis')->select('jadwal_terapis.*')->join('h_rekam_medis', 'jadwal_terapis.id_asses', '=', 'h_rekam_medis.id_asses')->where('h_rekam_medis.id_rm', $id)->whereNotIn('id_jadwal', DB::table('d_rekam_medis')->pluck('id_jadwal'))->where('status_terapis',  '=', 'Hadir')->get();
+            $detail = DetailRekamMedis::join('jadwal_terapis', 'd_rekam_medis.id_jadwal', '=', 'jadwal_terapis.id_jadwal')->where('id_rm', $id)->get();
+            return view('rekam_medis.detail_rekamedis', [
+                'detail' => $detail,
+                'id_rm' => $id,
+                'jadwal' => $jadwal
+            ]);
+        }else{
+            Alert::error('Silahkan login terlebih dahulu!')->autoclose(3500);
+            return redirect('/login');
+        }
     }
 
     /**
@@ -98,10 +111,16 @@ class detail_rekam_medis extends Controller
      */
     public function destroy(Request $r, $id)
     {
-        $hapus = DetailRekamMedis::where('id_sesirm', $id);
-        $hapus->delete();
+        if (Session::get('login')) {
+            
+            $hapus = DetailRekamMedis::where('id_sesirm', $id);
+            $hapus->delete();
 
-        Alert::success('Data berhasil dihapus')->autoclose(3500);
-        return redirect('detail_rekam_medis/'.$r->id_rm);
+            Alert::success('Data berhasil dihapus')->autoclose(3500);
+            return redirect('detail_rekam_medis/'.$r->id_rm);
+        }else{
+            Alert::error('Silahkan login terlebih dahulu!')->autoclose(3500);
+            return redirect('/login');
+        }
     }
 }
